Refactor view logic into its own module, PayDerbyDues::View
This module contains two functions: render and render_layout
to render a plain Handlebars template and to render onto and then
stick it into the layout.hbs template. This makes the code in
PayDerbyDues::Workflows::All much shorter and cleaner, and also
makes it easier to run the code outside EC2.