Add new user workflow
The basic idea is that the admin will fill in the new user's name and
email. This triggers an email to be sent to that user with a link.
This link contains an auth token and takes the user to a page that
lets them set up their account. There they can set a password and update
their info. Doing so successfully uses up the token from the link.

This commit also includes a bunch of smaller refactorings.